refactor: 통계 서비스 거래내역 서비스에서 분리 (#97)
