+++ api/services/transaction.ts
@@ -5,18 +5,15 @@ import { db } from 'models/db';
 import paymentService from './payment';
 
 import errorGenerator from 'utils/error-generator';
+import { getSideDate, parseTransactionByDate } from 'utils/date';
 
 import {
   PostTransactionParamType,
   EditTransactionParamType,
   getTransactionParamType,
   TransactionRecordType,
-  DayTransactionType,
-  CalendarStatisticsType,
   TransactionDataType,
 } from 'types/transaction';
-import { CategoryType } from 'types/common';
-import { CategoryStatisticsType } from 'types/statistics';
 
 /**
  * {
@@ -204,195 +201,9 @@ async function checkUserTransaction(userId: string, transactionId: string): Prom
 // TODO: 날짜 validation 추가
 // function checkValidDate(date: string): boolean {}
 
-//날짜 시작한날 끝날 구하기 - util로 이동
-function getSideDate(year: number, month: number): { startDate: Date; endDate: Date } {
-  const lastDate = new Date(year, month, 0).getDate();
-  return {
-    startDate: new Date(year, month - 1, 1),
-    endDate: new Date(year, month - 1, lastDate),
-  };
-}
-
-//거래내역 파싱 - util로 이동
-const parseTransactionByDate = (
-  transactions: Array<TransactionRecordType>,
-): Array<DayTransactionType> => {
-  const result: Array<DayTransactionType> = [];
-  let dayRecord: DayTransactionType = { date: '', transaction: [] };
-
-  transactions.forEach((record) => {
-    if (dayRecord.date === record.date) {
-      dayRecord.transaction?.push(record);
-      return;
-    }
-    if (dayRecord.date) {
-      result.push(dayRecord);
-      dayRecord = { date: '', transaction: [] };
-    }
-    dayRecord.date = record.date;
-    dayRecord.transaction = [record];
-  });
-
-  if (dayRecord.date) result.push(dayRecord);
-
-  return result;
-};
-
-async function getStatistics(
-  uid: string,
-  type: string,
-  year: string,
-  month: string,
-  category: string,
-): Promise<CategoryStatisticsType | number[] | void | CalendarStatisticsType> {
-  if (type === 'category') {
-    if (!year || !month) {
-      throw errorGenerator({
-        code: 'req/query-not-found',
-        message: 'Required query not found',
-      });
-    }
-    const currentMonthStart = `${year}-${month}`;
-    const currentMonthEnd = `${year}-${+month + 1}`;
-
-    const transactionSnapshot = await db.Transaction.findAll({
-      attributes: ['category', 'price'],
-      where: {
-        USERId: uid,
-        price: {
-          [Op.lte]: 0,
-        },
-        date: {
-          [Op.lt]: new Date(currentMonthEnd),
-          [Op.gte]: new Date(currentMonthStart),
-        },
-      },
-    });
-
-    const categoryStatistics: CategoryStatisticsType = {
-      life: 0,
-      food: 0,
-      transport: 0,
-      shop: 0,
-      health: 0,
-      culture: 0,
-      etc: 0,
-    };
-
-    transactionSnapshot.forEach((t) => {
-      const category: CategoryType = t.getDataValue('category');
-      const price: number = +t.getDataValue('price');
-
-      categoryStatistics[category] += price;
-    });
-
-    return categoryStatistics;
-  }
-
-  if (type === 'trend') {
-    if (!year || !category) {
-      throw errorGenerator({
-        code: 'req/query-not-found',
-        message: 'Required query not found',
-      });
-    }
-
-    const transactionSnapshot = await db.Transaction.findAll({
-      attributes: ['date', 'price'],
-      where: {
-        USERId: uid,
-        price: {
-          [Op.lte]: 0,
-        },
-        category,
-        date: {
-          [Op.lte]: new Date(`${year}-12-31`),
-          [Op.gte]: new Date(`${year}-1-1`),
-        },
-      },
-    });
-
-    const trendStatistics = new Array(12).fill(0);
-
-    transactionSnapshot.forEach((t) => {
-      const price = t.getDataValue('price');
-      const date = t.getDataValue('date');
-      const month = new Date(date).getMonth();
-
-      trendStatistics[month - 1] += +price;
-    });
-
-    return trendStatistics;
-  }
-  if (type === 'calendar') {
-    return await getCalendarStatistics(uid, year, month);
-  }
-  return;
-}
-
-//todo 반환값 바꾸기
-async function getCalendarStatistics(
-  uid: string,
-  year: string,
-  month: string,
-): Promise<CalendarStatisticsType> {
-  if (!year || !month) {
-    throw errorGenerator({
-      code: 'req/query-not-found',
-      message: 'Required query not found',
-    });
-  }
-  const result: CalendarStatisticsType = {
-    totalIncome: 0,
-    totalExpenditure: 0,
-    totalPrice: 0,
-    statistics: {},
-  };
-
-  const { startDate, endDate } = getSideDate(+year, +month);
-  const transactionSnapshot = await db.Transaction.findAll({
-    attributes: ['date', 'price'],
-    where: {
-      USERId: uid,
-      date: {
-        [Op.between]: [startDate, endDate],
-      },
-    },
-    order: [['date', 'DESC']],
-  });
-
-  transactionSnapshot.forEach((record) => {
-    const date = new Date(record.getDataValue('date')).getDate();
-    const price = +record.getDataValue('price');
-
-    if (price > 0) {
-      result.totalIncome += price;
-      if (date in result.statistics) {
-        result.statistics[date].income += price;
-        result.statistics[date].total += price;
-      } else {
-        result.statistics[date] = { income: price, expenditure: 0, total: price };
-      }
-    } else {
-      result.totalExpenditure += +price;
-      if (date in result.statistics) {
-        result.statistics[date].expenditure += price;
-        result.statistics[date].total += price;
-      } else {
-        result.statistics[date] = { income: 0, expenditure: price, total: price };
-      }
-    }
-  });
-
-  result.totalPrice = result.totalIncome + result.totalExpenditure;
-
-  return result;
-}
-
 export default {
   getTransaction,
   createTransaction,
   deleteTransaction,
   editTransaction,
-  getStatistics,
 };
